Trap block improvements.
The disguise portion of the trap block has been split to a separate class, which will make it easier to add other disguised blocks in the future.
Fixed flickering caused when right-clicking near the trap block.
Made it possible to disable the 3 different trigger methods of the trap block (stepping, interacting and breaking) separately.
Trap blocks now have the same hardness and blast resistance as the block it is disguised as.
Fixed trap blocks not dropping the disguised block's drop item when destroyed.