Cutscenes no longer affect entities in the real world.
Instead, they will be copied into the cutscene whenever their chunk is loaded by the cutscene. When this happens, the entities in the real world will be hidden from players in the cutscene.

Also load scoreboards before entity data in the cutscene to avoid a null pointer exception when loading display entities.