Removed the builtin TaskScheduler.
The TaskScheduler interface now instead points to Mojang's scheduler (and MinecraftServer#executeAsync in the case of scheduleImmediately).
Added a simple Throwaway schedule type to Mojang's scheduler which is never saved to disk. Meant for situations where it is not super important that the event runs, such as spawning particles or playing sounds.
Added a PlayerTeleport event which teleports a player even if they're offline when it runs. This event is now used for the Campus Lodestone.
Improved a variety of helpers and moved some cutscene and dungeon specific stuff to metacraft-lib.