Add soft reloads.
This reloads the config without updating resource packs for players immediately, but still gives the updated resource pack configuration to new players who log in.
Useful to silently remove resource packs after events for example, or if a smaller change was made that players don't need to see straight away.